+++ frontend/src/pages/podcast/PodcastDetailPage/PodcastDetailPage.tsx
@@ -4,7 +4,7 @@ import { useNavigate, useParams, useSearchParams } from "react-router"
 import { IoArrowBackSharp, IoReload } from "react-icons/io5"
 import LoadingDisplay from "../../../components/LoadingDisplay/LoadingDisplay.tsx"
 import { PodcastEpisode } from "../../../api/podcast/model/podcast.ts"
-import PodcastEpisodeCard from "../../../components/PodcastEpisodeCard/PodcastEpisodeCard.tsx"
+import PodcastEpisodeCard from "../../../components/PodcastEpisodeCard/index.tsx"
 import PodcastCard from "../../../components/PodcastCard/index.tsx"
 import { PodcastEpisodeContext } from "../../../context/PodcastEpisodeProvider/PodcastEpisodeProvider.tsx"
 import usePodcastEpisodes from "../../../hooks/podcast/usePodcastEpisodes.ts"
